Refactor index.js into modules
