+++ index.js
@@ -1,11 +1,12 @@
+import togglePages from "./modules/toggle.js";
+import addBookToLocalStorage from "./modules/addLocalStorage.js";
+import removeBook from "./modules/removeBook.js";
+import addBookToList from "./modules/addList.js";
+import displayBooks from "./modules/displayBooks.js";
+
 const form = document.querySelector('form');
 const bookList = document.querySelector('.book-list');
 const dateTime = document.querySelector('#date-time');
-const contacts = document.querySelector('.contacts');
-const bookForm = document.querySelector('.form');
-const listLink = document.querySelector('#nav__links__list');
-const addLink = document.querySelector('#nav__links__add');
-const contaLink = document.querySelector('#nav__links__contacts');
 
 const date = new Date();
 dateTime.innerHTML = date;
@@ -16,49 +17,6 @@ class Book {
     this.author = author;
     this.id = Math.random();
   }
-}
-
-function getBooks() {
-  let bookCollection;
-  if (localStorage.getItem('bookApp') === null) {
-    bookCollection = [];
-  } else {
-    bookCollection = JSON.parse(localStorage.getItem('bookApp'));
-  }
-  return bookCollection;
-}
-
-function addBookToLocalStorage(book) {
-  const bookCollection = getBooks();
-  bookCollection.push(book);
-  localStorage.setItem('bookApp', JSON.stringify(bookCollection));
-}
-
-function removeBook(id) {
-  const bookCollection = getBooks();
-  bookCollection.forEach((book, index) => {
-    if (book.id === id) {
-      bookCollection.splice(index, 1);
-    }
-    localStorage.setItem('bookApp', JSON.stringify(bookCollection));
-  });
-}
-
-function addBookToList(book) {
-  const newBook = document.createElement('div');
-  newBook.classList.add('oneBook');
-  newBook.innerHTML += `
-  <span hidden>${book.id}</span>
-  <tr>"${book.title}" by ${book.author}</tr>
-  <button type="button" class="remove">Remove</button>`;
-  bookList.appendChild(newBook);
-}
-
-function displayBooks() {
-  const books = getBooks();
-  books.forEach((book) => {
-    addBookToList(book);
-  });
 }
 
 document.addEventListener('DOMContentLoaded', displayBooks);
@@ -86,24 +44,5 @@ form.addEventListener('submit', (e) => {
   }
 });
 
-function togglePages(){
-  listLink.addEventListener('click', () => {
-    bookForm.classList.remove('enable');
-    contacts.classList.remove('enable');
-    bookList.classList.remove('disable');
-  });
-  
-  addLink.addEventListener('click', () => {
-    bookList.classList.add('disable');
-    contacts.classList.remove('enable');
-    bookForm.classList.add('enable');
-  });
-  
-  contaLink.addEventListener('click', () => {
-    bookList.classList.add('disable');
-    bookForm.classList.remove('enable');
-    contacts.classList.add('enable');
-  });
-}
 togglePages();
