+++ modules/getBook.js
@@ -1,0 +1,10 @@
+const getBooks = () => {
+  let bookCollection;
+  if (localStorage.getItem('bookApp') === null) {
+    bookCollection = [];
+  } else {
+    bookCollection = JSON.parse(localStorage.getItem('bookApp'));
+  }
+  return bookCollection;
+}
+export default getBooks;
